day_3/ai-first-day-3-activity-4/requirements.txt
openai==0.28.1
tiktoken==0.6.0
langchain==0.1.20
newspaper3k==0.2.8
streamlit
streamlit_option_menu
streamlit_extras
faiss-cpu
plotly
openpyxl
langchain_community
scipy
scikit-learn
lxml_html_clean